refactor(session): flatten session/default/* up one level
The 3-line session.rs / default.rs indirection was a tmux-era vestige; the
single-PTY shell session is the only session. Plugin renames follow in the
rename phase.
