+++ src/session/exit.rs
